Added an ingame-observer.yaml file
OpenRA PR 20436 removed the common file that none of the default mods use, so adding a copy of the one from TS here.
Removed redundant (and crashing) references to HPF and statistics widgets - the world actor doesn't have the necessary traits because it doesn't need them, so the widgets crash.